requirements.txt
streamlit>=1.32.0
deep-translator>=1.11.4
pypdf>=4.0.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
docx2python>=2.0.0
langdetect>=1.0.9
streamlit-option-menu>=0.3.6
python-magic>=0.4.27
flask>=2.3.3
werkzeug>=2.3.7
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
fpdf2>=2.7.6
